refactor content pipeline (part 2)
