+++ lib/posts.ts
@@ -1,8 +1,7 @@
 import * as fs from 'fs';
 import * as path from 'path';
-import matter from 'gray-matter';
-import { renderMarkdown } from './markdown-to-html';
 import { getContentDirectory } from './content';
+import { getPostBySlug } from './item-by-slug';
 import { PostKind } from '~/types/default';
 
 // Regex to parse date and title from the filename
@@ -12,41 +11,6 @@ const contentDirectory = '_content/posts';
 
 export function getPostPaths(category: PostKind = 'article') {
   return fs.readdirSync(path.join(getContentDirectory(contentDirectory), category));
-}
-
-export function getPostBySlug(
-  paths: string[],
-  fields: string[] = [],
-  category: PostKind = 'article'
-) {
-  const [year, month, day, slug] = paths;
-  const actualSlug = `${year}-${month}-${day}-${slug}.md`;
-  const contents = fs.readFileSync(
-    path.join(getContentDirectory(contentDirectory), category, actualSlug)
-  );
-  const { data, content } = matter(contents);
-
-  const items: any = {};
-
-  // Ensure only the minimal needed data is exposed
-  fields.forEach(field => {
-    if (field === 'slug') {
-      items[field] = `${year}/${month}/${day}/${slug}`;
-    }
-    if (field === 'content') {
-      items[field] = renderMarkdown(content || '');
-    }
-    // if no date set, generate from filename.
-    if (field === 'date' && !data[field]) {
-      items[field] = new Date(`${year}-${month}-${day}`).toISOString();
-    }
-
-    if (data[field]) {
-      items[field] = data[field];
-    }
-  });
-
-  return items;
 }
 
 export function getPostParams(slug: string): string[] {
